WIP using upstream bb (global or nix)

+++ barretenberg_wrapper/src/blake2s.rs
@@ -4,7 +4,7 @@ pub fn hash_to_field(input: &[u8]) -> [u8; 32] {
     let mut r = [0_u8; 32];
     let data = input.as_ptr() as *const u8;
     unsafe {
-        blake2s_to_field(data, input.len() as u64, r.as_mut_ptr());
+        blake2s_to_field(data, input.len(), r.as_mut_ptr());
     }
     r
 }
